fix: change datatype for setPenaltyNotRevealNum governance function

+++ contracts/Core/parameters/Governance.sol
@@ -64,7 +64,7 @@ contract Governance is Initializable, ACL, Constants {
      * @dev can be called only by the the address that has the governer role
      * @param _penaltyNotRevealNumerator updated value to be set for penaltyNotRevealNumerator
      */
-    function setPenaltyNotRevealNum(uint16 _penaltyNotRevealNumerator) external initialized onlyRole(GOVERNER_ROLE) {
+    function setPenaltyNotRevealNum(uint32 _penaltyNotRevealNumerator) external initialized onlyRole(GOVERNER_ROLE) {
         emit ParameterChanged(msg.sender, "penaltyNotRevealNum", _penaltyNotRevealNumerator, block.timestamp);
         rewardManagerParams.setPenaltyNotRevealNum(_penaltyNotRevealNumerator);
     }
